requirements.txt
# Unified Requirements for CLIP + DeepSORT + YOLOv5 Integration
# For NVIDIA GPU (CUDA 11.8+)

# PyTorch (install via pip with CUDA support)
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2

# YOLOv5 Dependencies
gitpython>=3.1.30
matplotlib>=3.3
numpy>=1.23.5,<2.0.0
opencv-python>=4.1.1
pillow>=10.3.0
psutil
PyYAML>=5.3.1
requests>=2.32.2
scipy>=1.4.1
thop>=0.1.1
tqdm>=4.66.3
ultralytics>=8.2.64
pandas>=1.1.4
seaborn>=0.11.0
packaging
setuptools>=70.0.0
urllib3>=2.5.0

# CLIP Dependencies
ftfy
regex

# DeepSORT Dependencies (GPU version)
tensorflow[and-cuda]==2.10.0
tf-slim
tf-keras

# Additional Integration Dependencies
scikit-learn
filterpy  # For Kalman filter

